LDAP frontend spike
A hacky proof-of-concept implementation of one way
that the CLI might be able to prompt the user for
their LDAP username and password and run an OIDC
authcode flow without opening a web browser.

The authorization endpoint has been changed to
accept basic auth credentials for LDAP IDPs.

The CLI has been changed to send the credentials
via basic auth and skip opening the web browser.
This implemenation uses the CLI's localhost port
listener to finish the flow, although that isn't
strictly required.

Signed-off-by: Andrew Keesler <akeesler@vmware.com>